+++ src/ResourceManager/Compute/Commands.Compute/Strategies/Network/NetworkSecurityGroupPolicy.cs
@@ -20,6 +20,7 @@ using System;
 using Microsoft.Azure.Commands.Common.Strategies;
 using Microsoft.Azure.Management.Compute.Models;
 using Microsoft.Azure.Commands.Compute.Strategies.ComputeRp;
+using System.Collections.Generic;
 
 namespace Microsoft.Azure.Commands.Compute.Strategies.Network
 {
@@ -36,14 +37,14 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.Network
                 createTime: _ => 15);
 
         public static ResourceConfig<NetworkSecurityGroup> CreateNetworkSecurityGroupConfig(
-            this ResourceConfig<ResourceGroup> resourceGroup, string name, Func<int[]> getOpenPorts)
+            this ResourceConfig<ResourceGroup> resourceGroup, string name, IList<int> openPorts)
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
                 name: name,
                 createModel: _ => new NetworkSecurityGroup
                 {
-                    SecurityRules = getOpenPorts()
-                        .Select((port, index) => new SecurityRule
+                    SecurityRules = openPorts
+                        ?.Select((port, index) => new SecurityRule
                         {
                             Name = name + port,
                             Protocol = "Tcp",
@@ -57,14 +58,5 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.Network
                         })
                         .ToList()
                 });
-
-        public static ResourceConfig<NetworkSecurityGroup> CreateNetworkSecurityGroupConfig(
-            this ResourceConfig<ResourceGroup> resourceGroup,
-            string name,
-            int[] openPorts,
-            Func<OperatingSystemTypes> getOsType)
-            => resourceGroup.CreateNetworkSecurityGroupConfig(
-                name,
-                () => getOsType().UpdatePorts(openPorts));
     }
 }