+++ src/ResourceManager/ManagementPartner/ChangeLog.md
@@ -1,4 +1,4 @@
-﻿<!--
+<!--
     Please leave this section at the top of the change log.
 
     Changes for the current release should go under the section titled "Current Release", and should adhere to the following format:
@@ -18,5 +18,11 @@
         - Additional information about change #1
 -->
 ## Current Release
-* First release of Management Partners cmdlets
 
+## Version 0.1.1
+* Add Partner Name in the response for cmdlets: Get-AzureRmManagementPartner, New-AzureRmManagementPartner and Update-AzureRmManagementPartner
+* Updated to the latest version of the Azure ClientRuntime
+* Set minimum dependency of module to PowerShell 5.0
+
+## Version 0.1.0
+* First release of Management Partners cmdlets